Add mechanism for head attacks
If our head is 2 moves away from another snake's head, we have the
possibility of killing that snake.  We assume that the other snake
will try to path to its tail and so we attempt to move into its head.
If the enemy head and tail is seperated, we calculate the shortest path
that the enemy snake may take to its tail and if we can move into the
next space that the enemy will, we do. We only consider spaces that both
us and the other snake can move to.